Use integers instead of floats for custom leaderboards

+++ DevilDaggersCore/CustomLeaderboards/CustomEntryBase.cs
@@ -6,7 +6,7 @@ namespace DevilDaggersCore.CustomLeaderboards
 	{
 		public int PlayerId { get; set; }
 		public string Username { get; set; }
-		public float Time { get; set; }
+		public int Time { get; set; }
 		public int Gems { get; set; }
 		public int Kills { get; set; }
 		public int DeathType { get; set; }
@@ -14,15 +14,15 @@ namespace DevilDaggersCore.CustomLeaderboards
 		public int ShotsFired { get; set; }
 		public int EnemiesAlive { get; set; }
 		public int Homing { get; set; }
-		public float LevelUpTime2 { get; set; }
-		public float LevelUpTime3 { get; set; }
-		public float LevelUpTime4 { get; set; }
+		public int LevelUpTime2 { get; set; }
+		public int LevelUpTime3 { get; set; }
+		public int LevelUpTime4 { get; set; }
 		public DateTime SubmitDate { get; set; }
 		public string ClientVersion { get; set; }
 
 		public double Accuracy => ShotsFired == 0 ? 0 : ShotsHit / (double)ShotsFired;
 
-		public CustomEntryBase(int playerId, string username, float time, int gems, int kills, int deathType, int shotsHit, int shotsFired, int enemiesAlive, int homing, float levelUpTime2, float levelUpTime3, float levelUpTime4, DateTime submitDate, string clientVersion)
+		public CustomEntryBase(int playerId, string username, int time, int gems, int kills, int deathType, int shotsHit, int shotsFired, int enemiesAlive, int homing, int levelUpTime2, int levelUpTime3, int levelUpTime4, DateTime submitDate, string clientVersion)
 		{
 			PlayerId = playerId;
 			Username = username;
